+++ lua/mini/pick.lua
@@ -16,7 +16,8 @@
 ---
 --- - Built-in pickers (see |MiniPick.builtin|):
 ---     - Files.
----     - Pattern match (for fixed pattern and with live feedback).
+---     - Pattern match (for fixed pattern or with live feedback; both allow
+---       file filtering via glob patterns).
 ---     - Buffers.
 ---     - Help tags.
 ---     - CLI output.
@@ -652,6 +653,10 @@
 
 ---@alias __pick_builtin_opts table|nil Options forwarded to |MiniPick.start()|.
 ---@alias __pick_builtin_local_opts table|nil Options defining behavior of this particular picker.
+---@alias __pick_builtin_grep_globs <globs> `(table)` - array of string glob patterns to restrict search to
+---     matching files. Supported only by "rg" and "git" tools, respects their
+---     specific glob syntax and effects. Default: `{}` (no restriction).
+---     Example: `{ '*.lua', 'lua/**' }` for Lua files and files in "lua" directory.
 
 ---@diagnostic disable:undefined-field
 ---@diagnostic disable:discard-returns
@@ -1304,12 +1309,15 @@ end
 ---     Default: whichever tool is present, trying in that same order.
 ---   - <pattern> `(string)` - string pattern to search. If not given, asks user
 ---     interactively with |input()|.
+---   - __pick_builtin_grep_globs
 ---@param opts __pick_builtin_opts
 MiniPick.builtin.grep = function(local_opts, opts)
-  local_opts = vim.tbl_deep_extend('force', { tool = nil, pattern = nil }, local_opts or {})
+  local_opts = vim.tbl_extend('force', { tool = nil, pattern = nil, globs = {} }, local_opts or {})
   local tool = local_opts.tool or H.grep_get_tool()
+  local globs = H.is_array_of(local_opts.globs, 'string') and local_opts.globs or {}
+  local name_suffix = #globs == 0 and '' or (' | ' .. table.concat(globs, ', '))
   local show = H.get_config().source.show or H.show_with_icons
-  local default_opts = { source = { name = string.format('Grep (%s)', tool), show = show } }
+  local default_opts = { source = { name = string.format('Grep (%s%s)', tool, name_suffix), show = show } }
   opts = vim.tbl_deep_extend('force', default_opts, opts or {})
 
   local pattern = type(local_opts.pattern) == 'string' and local_opts.pattern or vim.fn.input('Grep pattern: ')
@@ -1318,14 +1326,14 @@ MiniPick.builtin.grep = function(local_opts, opts)
     return MiniPick.start(opts)
   end
 
-  return MiniPick.builtin.cli({ command = H.grep_get_command(tool, pattern) }, opts)
+  return MiniPick.builtin.cli({ command = H.grep_get_command(tool, pattern, globs) }, opts)
 end
 
 --- Pick from pattern matches with live feedback
 ---
 --- Perform pattern matching treating prompt as pattern. Gives live feedback on
 --- which matches are found. Use |MiniPick-actions-refine| to revert to regular
---- matching.
+--- matching. Use `<C-o>` to restrict search to files matching glob patterns.
 --- Tries to use one of the CLI tools to create items (see |MiniPick-cli-tools|):
 --- `rg`, `git`. If none is present, error is thrown (for performance reasons).
 ---
@@ -1335,15 +1343,19 @@ end
 ---   Possible fields:
 ---   - <tool> `(string)` - which tool to use. One of "rg", "git".
 ---     Default: whichever tool is present, trying in that same order.
+---   - __pick_builtin_grep_globs
+---     Use `<C-o>` custom mapping to add glob to the array.
 ---@param opts __pick_builtin_opts
 MiniPick.builtin.grep_live = function(local_opts, opts)
-  local_opts = vim.tbl_deep_extend('force', { tool = nil }, local_opts or {})
+  local_opts = vim.tbl_extend('force', { tool = nil, globs = {} }, local_opts or {})
   local tool = local_opts.tool or H.grep_get_tool()
   if tool == 'fallback' or not H.is_executable(tool) then H.error('`grep_live` needs non-fallback executable tool.') end
 
+  local globs = H.is_array_of(local_opts.globs, 'string') and local_opts.globs or {}
+  local name_suffix = #globs == 0 and '' or (' | ' .. table.concat(globs, ', '))
   local show = H.get_config().source.show or H.show_with_icons
-  local default_opts = { source = { name = string.format('Grep live (%s)', tool), show = show } }
-  opts = vim.tbl_deep_extend('force', default_opts, opts or {})
+  local default_source = { name = string.format('Grep live (%s%s)', tool, name_suffix), show = show }
+  opts = vim.tbl_deep_extend('force', { source = default_source }, opts or {})
 
   local set_items_opts, spawn_opts = { do_match = false, querytick = H.querytick }, { cwd = opts.source.cwd }
   local process
@@ -1353,11 +1365,20 @@ MiniPick.builtin.grep_live = function(local_opts, opts)
     if #query == 0 then return MiniPick.set_picker_items({}, set_items_opts) end
 
     set_items_opts.querytick = H.querytick
-    local command = H.grep_get_command(tool, table.concat(query))
+    local command = H.grep_get_command(tool, table.concat(query), globs)
     process = MiniPick.set_picker_items_from_cli(command, { set_items_opts = set_items_opts, spawn_opts = spawn_opts })
   end
 
-  opts = vim.tbl_deep_extend('force', opts or {}, { source = { items = {}, match = match } })
+  local add_glob = function()
+    local ok, glob = pcall(vim.fn.input, 'Glob pattern: ')
+    if ok then table.insert(globs, glob) end
+    name_suffix = #globs == 0 and '' or (' | ' .. table.concat(globs, ', '))
+    MiniPick.set_picker_opts({ source = { name = string.format('Grep live (%s%s)', tool, name_suffix) } })
+    MiniPick.set_picker_query(MiniPick.get_picker_query())
+  end
+  local mappings = { add_glob = { char = '<C-o>', func = add_glob } }
+
+  opts = vim.tbl_deep_extend('force', opts or {}, { source = { items = {}, match = match }, mappings = mappings })
   return MiniPick.start(opts)
 end
 
@@ -3255,15 +3276,21 @@ H.grep_get_tool = function()
 end
 
 --stylua: ignore
-H.grep_get_command = function(tool, pattern)
+H.grep_get_command = function(tool, pattern, globs)
   if tool == 'rg' then
-    return {
-      'rg', '--column', '--line-number', '--no-heading', '--field-match-separator', '\\x00',
-      '--no-follow', '--color=never', '--', pattern
+    local res = {
+      'rg', '--column', '--line-number', '--no-heading', '--field-match-separator', '\\x00', '--no-follow', '--color=never'
     }
+    for _, g in ipairs(globs) do
+      table.insert(res, '--glob')
+      -- NOTE: no `*` as default is important to not "override" ignoring files
+      table.insert(res, g)
+    end
+    vim.list_extend(res, { '--', pattern })
+    return res
   end
   if tool == 'git' then
-    local res = { 'git', 'grep', '--column', '--line-number', '--null', '--color=never', '--', pattern }
+    local res = { 'git', 'grep', '--column', '--line-number', '--null', '--color=never', '-e', pattern, '--', unpack(globs) }
     if vim.o.ignorecase then table.insert(res, 6, '--ignore-case') end
     return res
   end